+++ src/sep.ira/calculator.maui/App.xaml.cs
@@ -1,0 +1,11 @@
+namespace cc.isr.Finance.Sep.Ira;
+
+public partial class App : Application
+{
+    public App() => this.InitializeComponent();
+
+    protected override Window CreateWindow( IActivationState? activationState )
+    {
+        return new Window( new AppShell() );
+    }
+}