+++ src/sep.ira/calculator.maui/AppShell.xaml.cs
@@ -1,0 +1,6 @@
+namespace cc.isr.Finance.Sep.Ira;
+
+public partial class AppShell : Shell
+{
+    public AppShell() => this.InitializeComponent();
+}