+++ src/sep.ira/calculator/docs/changelog.md
@@ -3,7 +3,10 @@
 All notable changes to this library will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/)
 
-[1.0.9663]: https://www.github.com/atecoder/dn.data.finance
+[1.0.9666]: https://www.github.com/atecoder/dn.data.finance
+
+## [1.0.9666] - 2026-06-19
+- Define financial values as Decimal.
 
 ## [1.0.9663] - 2026-06-16
 - Add report builder functionality.